Django/bookmarks/requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cloudinary==1.28.0
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.11
django-cloudinary-storage==0.3.0
django-extensions==3.1.3
easy-thumbnails==2.7.2
gunicorn==20.1.0
idna==3.3
oauthlib==3.1.1
Pillow==9.0.0
psycopg2==2.9.2
pycodestyle==2.8.0
pycparser==2.20
PyJWT==2.3.0
pyOpenSSL==21.0.0
python-dotenv==0.19.1
python3-openid==3.2.0
pytz==2021.3
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
Werkzeug==2.0.2
whitenoise==5.3.0
